Created Test Functions
Created test.c and test.h files to test the functionality of the
conversion functions. Doesn't include tests for the functions that print
to screen since those can't easily be tested in code and are more easily
tested by visual confirmation by the programmer.

Include test.h into the main file and call the test functions from
within main. All the functions have the form "uint32_t xx_Test(void)"
where xx is the name of the function it tests. All functions of this
form return an integer value indicating which case it failed on or if it
made it through all test cases correctly.