Lab 4 Report added and Lab 4&5 files updated
Updated the schematic for Lab 4 to reflect the final design change to
the circuitry between the tachometer output and the OPA2350 input.

Added files for Lab 4 report. Need to add data measurements and data
analysis and it'll be ready to turn in.

Doing initial add of Lab 5 files including the Lab 5 prep and Lab5
circuit.